# Speed up native assets target