--- a/packages/flutter_tools/lib/src/dart/package_map.dart
+++ b/packages/flutter_tools/lib/src/dart/package_map.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:typed_data';
+
 import 'package:package_config/package_config.dart';
 
 import '../base/common.dart';
@@ -26,7 +28,7 @@ Future<PackageConfig> loadPackageConfigWithLogging(File file, {
       if (!configFile.existsSync()) {
         return null;
       }
-      return configFile.readAsBytes();
+      return Future<Uint8List>.value(configFile.readAsBytesSync());
     },
     onError: (dynamic error) {
       if (!throwOnError) {