--- a/packages/flutter_tools/lib/src/test/test_compiler.dart
+++ b/packages/flutter_tools/lib/src/test/test_compiler.dart
@@ -168,7 +168,12 @@ class TestCompiler {
 
       Uri? nativeAssetsYaml;
       final Uri projectUri = FlutterProject.current().directory.uri;
-      final NativeAssetsBuildRunner buildRunner = NativeAssetsBuildRunnerImpl(projectUri, globals.fs, globals.logger);
+      final NativeAssetsBuildRunner buildRunner = NativeAssetsBuildRunnerImpl(
+        projectUri,
+        buildInfo.packageConfig,
+        globals.fs,
+        globals.logger,
+      );
       if (globals.platform.isMacOS) {
         (nativeAssetsYaml, _) = await buildNativeAssetsMacOS(
           buildMode: BuildMode.debug,